Change createClient to asynchronous createClientAsync API
One problem with synchronous createClient() is we don't necessarily know
which client code to use before receiving the ping response. Supposedly
it is safe to use Forge clients to connect to vanilla servers, so that's
not a very big problem. The bigger issue is that for servers that do not
respond to ping packets [1], the synchronous createClient() called
before the ping will still continue executing -- i.e., a race condition.

[1] https://github.com/PrismarineJS/node-minecraft-protocol/issues/327#issuecomment-174310946